# backend/swagger.yaml
openapi: 3.0.0
info:
  title: CoffeeMaker
  description: Node.js, mongoDB CoffeeMaker with login/registration system
  version: 1.0.0
servers:
  - url: http://localhost:5050
    description: Local development server
...
paths:
  /api/beans:
    get:
      tags:
      - Beans
      summary: List of beans
      description: User & Admin
      operationId: getBeans
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    [
                        {
                            "_id": "622d1841be2965a403e4d8e3",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Example",
                            "brand": "NAPA VALLEY COFFEE ROASTING COMPANY",
                            "roastProfile": "Medium",
                            "roastType": "Blend",
                            "priceKg": 170,
                            "origin": "Arabica",
                            "__v": 0
                        },
                        {
                            "_id": "622d1871be2965a403e4d8e5",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Indonesian Celebes Kalosi Vienna",
                            "brand": "MITALENA ARTISAN COFFEE",
                            "roastProfile": "Medium",
                            "roastType": "Single Origin",
                            "priceKg": 220,
                            "origin": "Arabica",
                            "__v": 0
                        },
                        {
                            "_id": "622d18aabe2965a403e4d8e7",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Peru FTO Decaf Dark",
                            "brand": "CAFE SANTANA",
                            "roastProfile": "Dark",
                            "roastType": "Single Origin",
                            "priceKg": 220,
                            "origin": "Organical",
                            "__v": 0
                        },
                        {
                            "_id": "622d18e2be2965a403e4d8e9",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Costa Rica Cordillera De Fuego Termico",
                            "brand": "THE CHOSEN BEAN",
                            "roastProfile": "Medium",
                            "roastType": "Single Origin",
                            "priceKg": 159,
                            "origin": "Arabica",
                            "__v": 0
                        },
                        {
                            "_id": "622d1917be2965a403e4d8eb",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Freight House",
                            "brand": "UTICA COFFEE ROASTING CO.",
                            "roastProfile": "Medium",
                            "roastType": "Blend",
                            "priceKg": 179,
                            "origin": "Arabica",
                            "__v": 0
                        },
                        {
                            "_id": "622d1957be2965a403e4d8ed",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Colombian Peaks",
                            "brand": "EIGHT O'CLOCK COFFEE COMPANY",
                            "roastProfile": "Medium",
                            "roastType": "Blend",
                            "priceKg": 219,
                            "origin": "Arabica",
                            "__v": 0
                        },
                        {
                            "_id": "622d1990be2965a403e4d8ef",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Dark Italian Roast",
                            "brand": "CRAVENS COFFEE COMPANY",
                            "roastProfile": "Dark",
                            "roastType": "Blend",
                            "priceKg": 143,
                            "origin": "Arabica",
                            "__v": 0
                        },
                        {
                            "_id": "622d1a5bbe2965a403e4d8f3",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "The 1950",
                            "brand": "IGNORE",
                            "roastProfile": "Medium",
                            "roastType": "Single Origin",
                            "processingMethod": "Washed",
                            "priceKg": 412,
                            "origin": "Arabica",
                            "__v": 0
                        },
                        {
                            "_id": "622d1a73be2965a403e4d8f6",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Buna Boka",
                            "brand": "JBC COFFEE ROASTERS",
                            "roastProfile": "Medium",
                            "roastType": "Single Origin",
                            "processingMethod": "Washed",
                            "priceKg": 302,
                            "origin": "Arabica",
                            "__v": 0
                        },
                        {
                            "_id": "622d2d8f1abd937c0c0cf9aa",
                            "beanAdded": "2022-05-20T19:25:22.018+00:00",
                            "typeOfBean": "Example",
                            "brand": "Example Company",
                            "roastProfile": "Dark",
                            "roastType": "Blend",
                            "priceKg": 192,
                            "origin": "Arabica",
                            "__v": 0
                        }
                    ]
        "400":
          description: Bean not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
    post:
      tags:
      - Beans
      summary: Add new bean
      description: Admin
      operationId: postBean
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                typeOfBean:
                  type: string
                  example: Evergood Classic
                brand:
                  type: string
                  example: Evergood
                roastProfile:
                  type: string
                  example: Medium
                roastType:
                  type: string
                  example: Blend
                priceKg:
                  type: number
                  example: 149
                origin:
                  type: string
                  example: Arabica
              required:
                - typeOfBean
                - brand
                - roastProfile
                - roastType
                - priceKg
                - origin
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "typeOfBean": "Evergood Classic",
                        "brand": "Evergood",
                        "roastProfile": "Medium",
                        "roastType": "Blend",
                        "priceKg": 149,
                        "origin": "Arabica",
                        "_id": "6288b238abd159f9448a1206",
                        "beanAdded": "2022-05-21T09:34:48.863Z",
                        "__v": 0
                    }
        "400":
          description: Please add all fields
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          description: Internal server error
  /api/beans/:id:
    get:
      tags:
      - Beans
      summary: Get specific bean
      description: Admin
      operationId: getBean
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "_id": "6288a4404571ebe68b8ac304",
                        "typeOfBean": "Evergood Classic",
                        "brand": "Evergood",
                        "roastProfile": "Medium",
                        "roastType": "Blend",
                        "priceKg": 149,
                        "origin": "Arabica",
                        "beanAdded": "2022-05-21T08:35:12.168Z",
                        "__v": 0
                    }
        "400":
          description: Bean not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
    delete:
      tags: 
      - Beans
      summary: Delete specific bean
      description: Admin
      operationId: deleteBean
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "id": "6288a4404571ebe68b8ac304"
                    }
        "400":
          description: Bean not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
    put:
      tags:
      - Beans
      summary: Update specific bean
      description: Admin
      operationId: updateBean
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                typeOfBean:
                  type: string
                  example: Evergood Classic
                brand:
                  type: string
                  example: Evergood
                roastProfile:
                  type: string
                  example: Medium
                roastType:
                  type: string
                  example: Blend
                priceKg:
                  type: number
                  example: 149
                origin:
                  type: string
                  example: Arabica
              required:
                - typeOfBean
                - brand
                - roastProfile
                - roastType
                - priceKg
                - origin
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "typeOfBean": "Evergood Classic",
                        "brand": "Evergood",
                        "roastProfile": "Medium",
                        "roastType": "Blend",
                        "priceKg": 149,
                        "origin": "Arabica",
                        "_id": "6288b238abd159f9448a1206",
                        "beanAdded": "2022-05-21T09:34:48.863Z",
                        "__v": 0
                    }
        "400":
          description: Bean not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
  /api/brews:
    get:
      tags: 
      - Brews
      summary: List of brews
      description: User & Admin
      operationId: getBrews
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "brewName": "My new brew",
                        "typeOfBean": "Evergood",
                        "gramsOfCoffee": 72,
                        "grindingSettings": 3,
                        "litersWater": 1.1,
                        "rating": 1,
                        "_id": "6288b3fddde43e0c28357b2d",
                        "brewedAt": "2022-05-21T09:42:21.873Z",
                        "__v": 0
                    }
        "400":
          description: Brew not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
    post:
      tags: 
      - Brews
      summary: Add new brew
      description: User & Admin
      operationId: postBrew
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brewName:
                  type: string
                  example: My new brew
                typeOfBean:
                  type: string
                  example: Evergood
                gramsOfCoffee:
                  type: number
                  example: 72
                grindingSettings:
                  type: number
                  example: 3
                litersWater:
                  type: number
                  example: 1.1
              required:
                - brewName
                - typeOfBean
                - gramsOfCoffee
                - grindingSettings
                - litersWater
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "brewName": "My new brew",
                        "typeOfBean": "Evergood",
                        "gramsOfCoffee": 72,
                        "grindingSettings": 3,
                        "litersWater": 1.1,
                        "rating": 1,
                        "_id": "6288b3fddde43e0c28357b2d",
                        "brewedAt": "2022-05-21T09:42:21.873Z",
                        "__v": 0
                    }
        "400":
          description: Please add all fields
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          description: Internal server error
  /api/brews/:id:
    get:
      tags: 
      - Brews
      summary: Get specific brew
      description: User & Admin
      operationId: getBrew
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "_id": "6288b3fddde43e0c28357b2d",
                        "brewName": "My new brew",
                        "typeOfBean": "Evergood",
                        "gramsOfCoffee": 72,
                        "grindingSettings": 3,
                        "litersWater": 1.1,
                        "rating": 3.6666666666666665,
                        "brewedAt": "2022-05-21T09:42:21.873Z",
                        "__v": 0
                    }
        "400":
          description: Brew not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
    delete:
      tags: 
      - Brews
      summary: Delete specific brew
      description: Admin
      operationId: deleteBrew
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "id": "6288b3fddde43e0c28357b2d"
                    }
        "400":
          description: Brew not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    put:
      tags: 
      - Brews
      summary: Update specific brew
      description: User & Admin
      operationId: updateBrew
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brewName:
                  type: string
                  example: My new brew
                typeOfBean:
                  type: string
                  example: Evergood
                gramsOfCoffee:
                  type: number
                  example: 72
                grindingSettings:
                  type: number
                  example: 3
                litersWater:
                  type: number
                  example: 1.1
              required:
                - brewName
                - typeOfBean
                - gramsOfCoffee
                - grindingSettings
                - litersWater
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "brewName": "My new brew",
                        "typeOfBean": "Evergood",
                        "gramsOfCoffee": 72,
                        "grindingSettings": 3,
                        "litersWater": 1.1,
                        "rating": 4.5,
                        "_id": "6288b3fddde43e0c28357b2d",
                        "brewedAt": "2022-05-21T09:42:21.873Z",
                        "__v": 0
                    }
        "400":
          description: Brew not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: >-
        Access token is missing or invalid, or the user does not have access to
        perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Not Found